More pythonic
Divide the program into function packages and a main program, and encapsulate the main program.
